refactor: 팀 UI 컴포넌트 분리
